almost is good

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: iharile <iharile@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/01 13:19:02 by iharile           #+#    #+#             */
-/*   Updated: 2021/12/02 20:46:55 by iharile          ###   ########.fr       */
+/*   Updated: 2021/12/03 10:13:42 by iharile          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,11 +26,11 @@ int	check_type(char c, va_list args)
 	if (c == 'x')
 		count = ft_hexa_lower(va_arg(args, int));
 	if (c == 'X')
-		count = ft_hexa_upper(va_arg(args, int));
+		count = ft_hexa_upper(va_arg(args, unsigned long));
 	if (c == 'u')
-		count = ft_unsigned_int(va_arg(args, int));
+		count = ft_unsigned_int(va_arg(args, unsigned long));
 	if (c == 'p')
-		count = ft_hex(va_arg(args, unsigned long long));
+		count = ft_hex(va_arg(args, unsigned long));
 	if (c == '%')
 		ft_putchar('%');
 	return (count);
